update losses and denom to float type

+++ src/comp0188_cw2/training/train_single_epoch.py
@@ -57,8 +57,8 @@ class TrainSingleEpoch:
             final loss for the epoch and a dictionary of predictions. The keys 
             will be the same keys required by the criterion. 
         """
-        losses = torch.tensor(0)
-        denom = torch.tensor(0)
+        losses = torch.tensor(0.0, dtype=torch.float32)
+        denom = torch.tensor(0.0, dtype=torch.float32)
         if gpu:
             _device = "cuda"
         else:
